Reduce the ext2 block size from u64 to u32.

+++ fal-backend-ext2/src/block_group.rs
@@ -16,10 +16,10 @@ impl BlockGroupDescriptor {
 }
 
 pub fn block_address(superblock: &superblock::Superblock, offset: u64) -> u32 {
-    (offset / superblock.block_size) as u32
+    (offset / u64::from(superblock.block_size)) as u32
 }
 pub fn block_offset(superblock: &superblock::Superblock, offset: u64) -> u32 {
-    (offset % superblock.block_size) as u32
+    (offset % u64::from(superblock.block_size)) as u32
 }
 pub fn load_block_group_descriptor<D: fal::Device>(
     filesystem: &Filesystem<D>,
@@ -70,13 +70,13 @@ pub fn inode_exists<D: fal::Device>(inode: u32, filesystem: &Filesystem<D>) -> i
 
     let bm_start_baddr = descriptor.inode_usage_bm_baddr;
     let bm_block_index =
-        u32::try_from(u64::from(index_inside_group / 8) / filesystem.superblock.block_size)
+        u32::try_from(u64::from(index_inside_group / 8) / u64::from(filesystem.superblock.block_size))
             .unwrap();
 
     let block_bytes = read_block(filesystem, bm_start_baddr + bm_block_index)?;
 
     let byte_index_inside_bm =
-        u32::try_from(u64::from(index_inside_group) / filesystem.superblock.block_size).unwrap();
+        u32::try_from(u64::from(index_inside_group) / u64::from(filesystem.superblock.block_size)).unwrap();
 
     let bm_byte = block_bytes[usize::try_from(byte_index_inside_bm).unwrap()];
     let bm_bit = 1 << ((inode - 1) % 8);
@@ -97,7 +97,7 @@ pub fn block_exists<D: fal::Device>(baddr: u32, filesystem: &Filesystem<D>) -> i
 
     let bm_start_baddr = descriptor.block_usage_bm_baddr;
     let bm_block_index =
-        u32::try_from(u64::from(index_inside_group / 8) / filesystem.superblock.block_size)
+        u32::try_from(u64::from(index_inside_group / 8) / u64::from(filesystem.superblock.block_size))
             .unwrap();
 
     let mut block_bytes = vec![0; usize::try_from(filesystem.superblock.block_size).unwrap()];
@@ -108,7 +108,7 @@ pub fn block_exists<D: fal::Device>(baddr: u32, filesystem: &Filesystem<D>) -> i
     )?;
 
     let byte_index_inside_bm =
-        u32::try_from(u64::from(index_inside_group) / filesystem.superblock.block_size).unwrap();
+        u32::try_from(u64::from(index_inside_group) / u64::from(filesystem.superblock.block_size)).unwrap();
 
     let bm_byte = block_bytes[usize::try_from(byte_index_inside_bm).unwrap()];
     let bm_bit = 1 << (baddr % 8);
